V.0.1 changes

+++ src/Controller/Dashboard/MarketPlace/Market/BrandController.php
@@ -2,11 +2,8 @@
 
 namespace App\Controller\Dashboard\MarketPlace\Market;
 
-use App\Entity\MarketPlace\Market;
 use App\Entity\MarketPlace\MarketBrand;
-use App\Form\Type\Dashboard\MarketPlace\ProductType;
 use App\Form\Type\Dashboard\MarketPlace\BrandType;
-use App\Security\Voter\ProductVoter;
 use App\Service\MarketPlace\MarketTrait;
 use App\Service\Dashboard;
 use Doctrine\ORM\EntityManagerInterface;
@@ -18,7 +15,6 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/dashboard/market-place/brand')]
@@ -151,7 +147,7 @@ class BrandController extends AbstractController
     {
         $market = $this->market($request, $user, $em);
 
-        if ($this->isCsrfTokenValid('delete', $request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete', $request->get('_token')) && !$brand->getMarketProductBrands()->count()) {
             $em->remove($brand);
             $em->flush();
         }
